+++ chat/preview/assets/browser.processor.raw.js
@@ -73,7 +73,7 @@ class FListImagePreviewDomMutator {
 
         const img = selected.filter(el => (el !== body)).shift();
 
-        this.debug('detectImage.found', img);
+        this.debug('detectImage.found', !!img, img);
 
         return img;
     }